Change single job endpoint to stream
This required dealing with the resource not existing. I've done this
simply by checking whether the state atom exists and pushing out a 404
if it is nil. The FakeChannel required a bit of poking as when you write
a map to a websocket it just takes :body and outputs it.