Convert /queue to stream
The UnboundedQueue needs to implement IRef so that we can attach
watchers to it. Whilst doing this I switched the underlying data
structure to something thread safe as I am sure multiple threads will be
adding to this buffer.

The queue needs to be kept in order so I have pushed through changes to
diff so that even if the old/new are lists we can force it through
clojure.data/diff which will keep order.